// src/lib/utils/types/documents.ts
export type Documents = {
  proofOfOwnership: Array<File>;
  proofOfBilling: Array<File>;
  barangayCertificate: Array<File>;
  validId: Array<File>;
  selfieWithValidId: Array<File>;

  proofOfOwnershipToUpload: Array<File>;
  proofOfBillingToUpload: Array<File>;
  barangayCertificateToUpload: Array<File>;
  validIdToUpload: Array<File>;
  selfieWithValidIdToUpload: Array<File>;
};